requirements.txt
Faker==26.2.0
allure-pytest==2.13.5
loguru==0.7.2
selenium==4.23.1
pytest==8.3.2
